+++ src/main.rs
@@ -7,6 +7,7 @@
 extern crate serde_derive;
 extern crate serde;
 extern crate serde_yaml;
+extern crate serde_json;
 
 extern crate futures;
 extern crate futures_cpupool;